consolidate serialization helpers into events.py
No cross-package consumers. Move SerializableException and
SerializableEventType (plus their serialize/deserialize helpers) into
workflows.events alongside SerializableEvent, and delete the separate
serialization_helpers module.
