+++ packages/llama-index-workflows/tests/runtime/test_tick_serialization.py
@@ -13,7 +13,11 @@ from workflows.events import (
     StartEvent,
     StopEvent,
     _deserialize_event,
+    _deserialize_event_type,
+    _deserialize_exception,
     _serialize_event,
+    _serialize_event_type,
+    _serialize_exception,
 )
 from workflows.runtime.types.results import (
     AddCollectedEvent,
@@ -22,12 +26,6 @@ from workflows.runtime.types.results import (
     DeleteWaiter,
     StepWorkerFailed,
     StepWorkerResult,
-)
-from workflows.runtime.types.serialization_helpers import (
-    _deserialize_event_type,
-    _deserialize_exception,
-    _serialize_event_type,
-    _serialize_exception,
 )
 from workflows.runtime.types.ticks import (
     TickAddEvent,
